backend/pyproject.toml
[project]
name = "prelegal-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.0",
    "fastapi>=0.115.0",
    "litellm>=1.83.14",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.12.5",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
